Implement the `parallel` and the `skip` flags
`xliff` downloads the xliff versions of the files
`--parallel` controls how many files will be downloaded at the same time.
`--skip` normally the client waits for all downloads to be successful
before updating the files on the disk. With this flag it ignores errors
and saves all succesful downloads immediately.