Include `minimum_perc` in download flow
This adds into `shouldSkipDownload` method the ability to cancel a
translation download based on the `minimum_perc` the user has added.

There are two ways this can be achieved:
a) By adding in `config` for each resource a `minumum_perc` or by using
a flag `minumum_perc` when using the `tx pull command`. The pull command
argument will always surpass the individual resource's setting.

During the implementation the `MinimumPercentage` was added for the
`config` library and is now taken under consideration for the migration
process.