Add ability to tx to run in a subfolder
Up until now, running `tx` was functioning with the assumption that was
run from the project root folder and that config was in this path.

With this change we enable someone to use `tx` from a subfolder of this
project and it will try recursively try to find the `.tx/config` file
until it hits a `/` or `.` based on `.Dir()` implementation.