Fix issue with paths on action from different OS
This solves a case when a user tries to make an action (pull/push) from
an OS that has different path separator from the one that was set in
`config`.

Example:

`config`

```
...
file_filter = src/static/_locales/<lang>/messages.json
...
```

When the user tries to push or pull from a Windows machine, this will
fail due to the way we search the `searchFileFilter` recursively
searches for files (`os.PathSeparator`).

The current fix tries to find if the `os.PathSeparator` is contained
in the path and replaces with the alternative.

In the specific example we will search for `\` which is the OS separator
and since there is no occurrence we will replace `/` with `\` and we
will be able to locate the files. This should also apply when we have a
Windows `file_filter` and a Linux based machine that tries to make an
action.

An additional message that we didn't find any translation files was
added to provide more context to the user.