Fix issue when testing empty dir exists
When pulling, we check if the directory of the source file exists.
In case the path is the current directory, this fails and returns that
directory doesn't exist.

Example:

```
source_file  = cli.txt
```

`filepath.Split(cfgResource.SourceFile)` will return an empty string
and `if _, statErr := os.Stat(dir); os.IsNotExist(statErr)` will return
an error that the `directory '' does not exist`.

This fix will check first if the `dir` is not empty (current directory)
and then proceed to check if it exists.