Fix migrate issue with SourceFile
There seems to be cases that source_file is omitted from `config` and
the combination of `file_filter` and `source_language` was used.

In order not to break users' migrations, during the migration script we
check if the `source_file` exists and if not we try to create it with
the `file_filter` and `source_language`.