Fix issue that "-f" affects "minimum-perc"
Minimum percentage checks should not be affected by the force flag.

For this to be solved in order to get to the "shouldSkipDownload"
method, we take under consideration also the minimum-perc and
"shouldSkipDownload" also is aware of the force flag in order to decide
which checks it will make.

If minimum-perc exists and fails the checks, then the pull for the
specific resource will fail, even if force flag is true.

Also `CreateLanguageDirectory` method was renamed to
`GetLanguageDirectory` to better reflect its functionality.