Fix issue with tar inconsistency
This introduces a condition in the installation script that tries to
identify if the tar implementation is GNU.
If not, the script fallbacks to a `bsdtar` compatible command.

The reason behind this change is that `--skip-old-files` is not
available in `bsdtar` but a `-k, --keep-old-files` which has the same
result is.