Add download pseudo translations flag
This uses the `pseudo` attribute in API to download
files with pseudo translations.

The change is on the pull command, in which, if
`pseudo` flag is true we use the source file
download instead of the translations file download
because for compatibility issues. The `pseudo`
attribute will be removed from the API for translations
in the future and will be used only in source.