Fix issue with minimum-perc skip message
Before the fix, we were returning a message that a local file is newer
than the remote one, which wasn't true.

This fix updates the `shouldSkipDownload` to return an additional value
with a message in case it's different than the default one.

Also fixed a typo in readme regarding the `minimum-perc` flag.